Give extra credit if find match within bonus time limit.

+++ Memo/EmojiMemoryGameViewModel.swift
@@ -26,7 +26,7 @@ class EmojiMemoryGameViewModel: ObservableObject {
         model.cards
     }
     
-    var score: Int {
+    var score: Double {
         model.score
     }
